superhero
This is a website where you can search for any fictional character . The relevant information regarding the character is displayed. You can also add your favourite characters to a Favourites list, and access them from the 'Favourites' section.

Home Page: Type the name of the character you want to search, in the given search bar. Real-time suggestions are provided.

Superhero page: This page contains all the information regarding a character. You can see the image of the character, and its biography, power stats, appearance, work, and connections. The 'heart icon' below the image lets you add the character to your Favourites list, or remove it from Favourites if its already added.

Favourites page: This provides a persistent list of all the characters you Liked. You can remove a character from the Favourites list directly from this page by clicking the 'heart icon'. Clicking on the image of any of the characters redirects you to the Superhero page of that character.